refactor: move the function to a separate file
